Remove unused provider and rename MsJwtProvider to JwtProvider
